+++ CHANGELOG.md
@@ -1,12 +1,15 @@
 0.6.2
 ----
 * Added `announce`, `announceMaybe` and `announceMaybeT` which invoke a method for every objects in a container.
+* Added `(@**@)` and `(@||@)`.
 * Renamed `(.>>.)` to `(@>>@)`, `.|>.` to `(@|>@)` for consistency.
 * Added `filterPush`.
 * Added `iterObject`, `iterTObject`, `iterative`, `iterativeT` for free monads.
 * Renamed `runSequential` to `(@!)`.
 * Added combinators for `ReifiedProgramT`: `(@!!)` and `sequentialT`.
 * Changed the semantics of `variable` to accept `StateT`.
+* Added `flyweight'` that relies on HashMap
+* Added `MonadObjective` constraint
 
 0.6.1
 -----